Reorganize ne_NP address test
